Restrict motor outs and sensor ins at HMM level
Do it *properly* this time. This commit fixes another horrible bug with
modding... uses the actual mod operator, rather than &, so it doesn't
need to be modding by some (2^n - 1), since now the range of in/out
nodes is not of that form. Also, we were subtracting by 1 when we
shouldn't have been.